pyproject.toml
[tool.poetry]
name = "nonebot-plugin-group-fuukiiin"
version = "0.1.0"
description = "用来监察群友干坏事(重复刷屏，违禁词，违禁表情包)的群风纪委员!"
authors = ["Azide <rukuy@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "nonebot_plugin_group_fuukiiin", from = "./src/plugins/" }
]

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = ">=2.0.0-rc.2"
nb-cli = "^0.6.7"
transitions = "^0.9.0"
pyenchant = "^3.2.2"
pytest = "^7.2.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true

[tool.nonebot]
plugins = ["src.plugins.nonebot_plugin_group_fuukiiin"]